% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrientMode.R
\name{OrientMode}
\alias{OrientMode}
\title{Find the statistical mode of a vector.}
\usage{
OrientMode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The statistical mode of \code{x} as ox and the corresponding stress 
orientation as sx.
}
\description{
Find the statistical mode of a vector.
}
\examples{
OrientMode(1, 1, 1, 3, 4, 5, 2)
}
