% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrientRose.R
\name{OrientRose}
\alias{OrientRose}
\title{Build a rose diagram from a data frame column.}
\usage{
OrientRose(df, f)
}
\arguments{
\item{df}{A data frame.}

\item{f}{A column in the data frame.}
}
\value{
A rose diagram with lines showing preferred orientation and stress
orientation.
}
\description{
Build a rose diagram from a data frame column.
}
\examples{
OrientRose(10, 10, 10, 300, 140, 50, 260)
}
